Moved app module components to app layer, made domain pure kotlin module
