+++ app/src/main/java/com/arch/template/base/BaseActivity.kt
@@ -7,8 +7,8 @@ import androidx.databinding.DataBindingUtil
 import androidx.databinding.ViewDataBinding
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
-import com.arch.R
 import com.arch.presentation.viewmodels.base.BaseViewModel
+import com.arch.template.R
 
 abstract class BaseActivity<B : ViewDataBinding, VM : BaseViewModel> : AppCompatActivity() {
