+++ app/src/main/java/com/arch/template/utils/ResourceString.kt
