+++ app/src/main/java/com/arch/template/utils/extension/DimensionExtension.kt
@@ -1,4 +1,4 @@
-package com.arch.utils.extensions
+package com.arch.template.utils.extension
 
 import android.content.res.Resources
