+++ domain/build.gradle
@@ -1,59 +1,12 @@
 plugins {
-    id 'com.android.library'
-    id 'kotlin-android'
+    id 'java-library'
+    id 'kotlin'
     id 'kotlin-kapt'
 }
-apply from: "../jacoco.gradle"
 
-android {
-    compileSdk AppConfig.compileSdk
-
-    defaultConfig {
-        minSdk AppConfig.minSdk
-        targetSdk AppConfig.targetSdk
-
-        testInstrumentationRunner AppConfig.androidTestInstrumentation
-        consumerProguardFiles "consumer-rules.pro"
-    }
-
-    buildTypes {
-        debug {
-            testCoverageEnabled true
-        }
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
-        }
-        dev {
-            initWith buildTypes.release
-        }
-
-        qa {
-            initWith buildTypes.release
-        }
-
-        staging {
-            initWith buildTypes.release
-        }
-
-        prod {
-            initWith buildTypes.release
-        }
-
-    }
-    compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
-    }
-    buildFeatures {
-        dataBinding true
-    }
-    buildFeatures {
-        viewBinding true
-    }
-    kotlinOptions {
-        jvmTarget = '1.8'
-    }
+java {
+    sourceCompatibility = JavaVersion.VERSION_1_8
+    targetCompatibility = JavaVersion.VERSION_1_8
 }
 
 dependencies {
@@ -61,20 +14,20 @@ dependencies {
     implementation fileTree(dir: "libs", include: ["*.jar"])
     implementation project(path: ':utils')
 
-    //app libs
-    implementation(AppDependencies.appLibraries)
-    implementation AppDependencies.hiltAndroid
+
+    implementation AppDependencies.hiltCore
+    implementation AppDependencies.javaxInject
+
+    //kpat libs
+    kapt AppDependencies.hiltCoreCompiler
+    kapt AppDependencies.hiltAndroidCompiler
+
     //paging
-    implementation AppDependencies.paging
+    implementation AppDependencies.pagingCoreCommon
 
     //test libs
     testImplementation(AppDependencies.testLibraries)
-    //kpat libs
-    kapt(AppDependencies.kpatLibraries)
-    androidTestImplementation(AppDependencies.androidTestLibraries)
-}
 
-// Allow references to generated code
-kapt {
-    correctErrorTypes = true
+
+
 }
