+++ domain/src/main/java/com/arch/utils/Validator.kt
@@ -1,5 +1,7 @@
 package com.arch.utils
 
+import com.arch.utils.extension.isValidEmail
+
 class Validator {
     companion object Factory {
         fun isEmpty(value: String): Boolean {
