+++ domain/src/main/java/com/arch/utils/extension/StringExtensions.kt
@@ -1,4 +1,4 @@
-package com.arch.utils
+package com.arch.utils.extension
 
 import java.util.regex.Pattern
